Framgang??????
So, the phong shaded house is showing up, but it has the dog texture.
Although not *really* a problem, but unsure whether or not phong shader is actually applied. The house has a yellow shade to it, which is probably the "light" doing its work, but it does not seem to be affected by the light's strength...